Move client.rs to common dependency
