+++ google-apis-common/Cargo.toml
@@ -26,4 +26,5 @@ http = "^0.2"
 tokio = "^1.0"
 tower-service = "^0.3.1"
 
-
+[dev-dependencies]
+serde = { version = "^ 1.0", features = ["derive"] }
