--- src/rust/api/client.rs
@@ -1,2 +1,0 @@
-//! This file serves on purpose and is an artifact of the buildsystem.
-//! Its content can now be found in the `google-apis-common` crate.
